Crossbow + Player Movement test
Crossbow is added, and a test for a dodgeroll has been added but doesnt quite work.

+++ Assets/Player/Player_Scripts/PlayerCore.cs
@@ -2,10 +2,10 @@ using UnityEngine;
 
 public class Player : MonoBehaviour
 {
-    [SerializeField] private int maxHealth = 100;
-    [SerializeField] private int currentHealth = 100;
+    [SerializeField] private float maxHealth = 100f;
+    [SerializeField] private float currentHealth = 100f;
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         currentHealth -= damage;
         Debug.Log($"Player took {damage} damage. Remaining HP: {currentHealth}");
